MainPage -> LoginPage 수정 styling 파일 별도로 분리 [파일이름].style.ts theme color같이 global하게 쓸수있는 폴더 생성 (util 함수나 mixin같은거 추후 추가 예정)
